Conflicts:
	cmake/compilers/CXXFlags.cmake

Former-commit-id: d89a3d7826f9d2f82bd6394b44fbb1216b49ef04

+++ src/lib/libmoinfo/model_space_build.cc
@@ -91,8 +91,8 @@ void ModelSpace::build()
   if(beta_combinations.size()==0)
     beta_combinations.push_back(std::vector<int>(0));
 
-  for(int a = 0; a < alfa_combinations.size(); ++a){
-    for(int b = 0; b < beta_combinations.size(); ++b){
+  for(size_t a = 0; a < alfa_combinations.size(); ++a){
+    for(size_t b = 0; b < beta_combinations.size(); ++b){
       int alfa_sym = 0; // Symmetry of the alfa string
       int beta_sym = 0; // Symmetry of the beta string
 
@@ -136,7 +136,7 @@ void ModelSpace::build()
 
 void ModelSpace::classify()
 {
-  for(int mu = 0; mu < determinants.size(); ++mu){
+  for(size_t mu = 0; mu < determinants.size(); ++mu){
     if(determinants[mu].is_closed_shell()){
       closed_to_all.push_back(mu);
     }else{
